+++ src/main/java/org/glycoinfo/application/glycanbuilder/util/canvas/CanvasActionDescriptor.java
@@ -3,7 +3,6 @@ package org.glycoinfo.application.glycanbuilder.util.canvas;
 public enum CanvasActionDescriptor {
 
 	STRINGNOTATION("stringNotation"),
-	ANTENNAPARENT("antennaParent"),
 	UNDO("undo"),
 	REDO("redo"),
 	CUT("cut"),